Add missing dependency
